* Animated Transitions: Seamless animations enhance the user experience when the dialog appears and disappears. * Customizable Content: Tailored messages can be displayed through the text and body properties, allowing for clear and concise communication with users. * Auto-Dismissal: Upon user interaction with the provided buttons, the dialog gracefully dismisses itself, ensuring a smooth interaction flow. * Clear Naming Convention: The show method clearly communicates the intention of displaying the dialog. * Optional Shadow Box: Implement a subtle shadow effect for added depth and visual hierarchy using the showWithShadow method (if applicable). * Cascading Alert Support: The provided example demonstrates the capability to chain multiple dialogs for complex workflows (if applicable). * Refactored Structure: The code is reorganized for improved readability, maintainability, and adherence to best practices. * Concise Imports: All necessary imports are consolidated into a single line for cleaner code. Additional Refinements: Minor improvements have been made to further polish the functionality and overall quality. * Type changed: from Delete to Caution
