--- .vscode/settings.json
@@ -1,22 +1,0 @@
-{
-    "workbench.colorCustomizations": {
-        "activityBar.activeBackground": "#098695",
-        "activityBar.background": "#098695",
-        "activityBar.foreground": "#e7e7e7",
-        "activityBar.inactiveForeground": "#e7e7e799",
-        "activityBarBadge.background": "#78076c",
-        "activityBarBadge.foreground": "#e7e7e7",
-        "commandCenter.border": "#e7e7e799",
-        "sash.hoverBorder": "#098695",
-        "statusBar.background": "#065b65",
-        "statusBar.foreground": "#e7e7e7",
-        "statusBarItem.hoverBackground": "#098695",
-        "statusBarItem.remoteBackground": "#065b65",
-        "statusBarItem.remoteForeground": "#e7e7e7",
-        "titleBar.activeBackground": "#065b65",
-        "titleBar.activeForeground": "#e7e7e7",
-        "titleBar.inactiveBackground": "#065b6599",
-        "titleBar.inactiveForeground": "#e7e7e799"
-    },
-    "peacock.color": "#065b65"
-}
