+++ lib/resources/colors.dart
@@ -4,3 +4,4 @@ const Color infoColor = Color(0xff0066FF);
 const Color multiActionColor = Color(0xfffb8500);
 const Color errorColor = Color(0xffFE355C);
 const Color permissionColor = Color(0xff6c757d);
+const Color dialogShadowColorConst = Color.fromRGBO(0, 0, 0, 0.55);
