+++ lib/widgets/dialogs_widgets/delete_elegant_alert.dart
@@ -1,5 +1,5 @@
 import 'package:elegant_alert_dialog/resources/colors.dart';
-import 'package:elegant_alert_dialog/resources/elegant_notification_type.dart';
+import 'package:elegant_alert_dialog/widgets/support_widgets/elegant_notification_type.dart';
 import 'package:flutter/material.dart';
 
 class DeleteElegantAlert extends StatelessWidget {
@@ -33,12 +33,13 @@ class DeleteElegantAlert extends StatelessWidget {
             const SizedBox(
               width: 10,
             ),
-            const Text(
-              'This is serious!',
-              style: TextStyle(
-                fontWeight: FontWeight.bold,
-              ),
-            ),
+            elegantAlertDeleteType.body?.getTitle() ??
+                const Text(
+                  'This is serious!',
+                  style: TextStyle(
+                    fontWeight: FontWeight.bold,
+                  ),
+                ),
           ],
         ),
         Expanded(
@@ -48,21 +49,36 @@ class DeleteElegantAlert extends StatelessWidget {
             ),
             child: Column(
               children: [
-                Text(
-                  'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.',
-                ),
+                elegantAlertDeleteType.body?.getBodyText() ??
+                    const Text(
+                      'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.',
+                    ),
                 Expanded(
                   child: Row(
                     mainAxisAlignment: MainAxisAlignment.end,
                     children: [
                       TextButton(
-                        onPressed: () {},
+                        onPressed: () {
+                          if (elegantAlertDeleteType.onCancelPressed != null) {
+                            elegantAlertDeleteType.onCancelPressed!();
+                            Navigator.of(context).pop();
+                          } else {
+                            Navigator.of(context).pop();
+                          }
+                        },
                         child: Text(
                           elegantAlertDeleteType.cancelButtonText,
                         ),
                       ),
                       ElevatedButton(
-                        onPressed: () {},
+                        onPressed: () {
+                          if (elegantAlertDeleteType.onDeletePressed != null) {
+                            elegantAlertDeleteType.onDeletePressed!();
+                            Navigator.of(context).pop();
+                          } else {
+                            Navigator.of(context).pop();
+                          }
+                        },
                         child: Text(
                           elegantAlertDeleteType.deleteButtonText,
                         ),
