+++ lib/widgets/dialogs_widgets/permission_elegant_alert.dart
@@ -1,5 +1,5 @@
 import 'package:elegant_alert_dialog/resources/colors.dart';
-import 'package:elegant_alert_dialog/resources/elegant_notification_type.dart';
+import 'package:elegant_alert_dialog/widgets/support_widgets/elegant_notification_type.dart';
 import 'package:flutter/material.dart';
 
 class PermissonElegantAlert extends StatelessWidget {
@@ -33,12 +33,13 @@ class PermissonElegantAlert extends StatelessWidget {
             const SizedBox(
               width: 10,
             ),
-            const Text(
-              'This is serious!',
-              style: TextStyle(
-                fontWeight: FontWeight.bold,
-              ),
-            ),
+            elegantAlertPermission.body?.getTitle() ??
+                const Text(
+                  'This is serious!',
+                  style: TextStyle(
+                    fontWeight: FontWeight.bold,
+                  ),
+                ),
           ],
         ),
         Expanded(
@@ -48,20 +49,37 @@ class PermissonElegantAlert extends StatelessWidget {
             ),
             child: Column(
               children: [
-                Text(
-                  'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.',
-                ),
+                elegantAlertPermission.body?.getBodyText() ??
+                    const Text(
+                      'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.',
+                    ),
                 Expanded(
                   child: Row(
                     mainAxisAlignment: MainAxisAlignment.end,
                     children: [
                       //TODO refactor buttons with other alert types
                       TextButton(
-                        onPressed: () {},
+                        onPressed: () {
+                          if (elegantAlertPermission.onDenyButtonPressed !=
+                              null) {
+                            elegantAlertPermission.onDenyButtonPressed!();
+                            Navigator.of(context).pop();
+                          } else {
+                            Navigator.of(context).pop();
+                          }
+                        },
                         child: Text(elegantAlertPermission.denyButtonText),
                       ),
                       ElevatedButton(
-                        onPressed: () {},
+                        onPressed: () {
+                          if (elegantAlertPermission.onAllowButtonPressed !=
+                              null) {
+                            elegantAlertPermission.onAllowButtonPressed!();
+                            Navigator.of(context).pop();
+                          } else {
+                            Navigator.of(context).pop();
+                          }
+                        },
                         child: Text(elegantAlertPermission.allowButtonText),
                       ),
                     ],
