+++ pubspec.yaml
@@ -3,7 +3,7 @@ description: A new Flutter package project.
 version: 0.0.1
 homepage: test
 environment:
-  sdk: '>=3.4.3 <4.0.0'
+  sdk: '>=3.4.1 <4.0.0'
   flutter: '>=1.17.0'
 dependencies:
   flutter:
